feat: do not scan not watched folders
