+++ src/context.rs
@@ -1000,12 +1000,6 @@ impl Context {
                 .to_string(),
         );
         res.insert(
-            "scan_all_folders_debounce_secs",
-            self.get_config_int(Config::ScanAllFoldersDebounceSecs)
-                .await?
-                .to_string(),
-        );
-        res.insert(
             "quota_exceeding",
             self.get_config_int(Config::QuotaExceeding)
                 .await?
