+++ src/scheduler.rs
@@ -530,38 +530,6 @@ async fn fetch_idle(
         .await
         .context("download_msgs")?;
 
-    // Scan additional folders only after finishing fetching the watched folder.
-    //
-    // On iOS the application has strictly limited time to work in background, so we may not
-    // be able to scan all folders before time is up if there are many of them.
-    if folder_config == Config::ConfiguredInboxFolder {
-        // Only scan on the Inbox thread in order to prevent parallel scans, which might lead to duplicate messages
-        match connection
-            .scan_folders(ctx, &mut session)
-            .await
-            .context("scan_folders")
-        {
-            Err(err) => {
-                // Don't reconnect, if there is a problem with the connection we will realize this when IDLEing
-                // but maybe just one folder can't be selected or something
-                warn!(ctx, "{:#}", err);
-            }
-            Ok(true) => {
-                // Fetch the watched folder again in case scanning other folder moved messages
-                // there.
-                //
-                // In most cases this will select the watched folder and return because there are
-                // no new messages. We want to select the watched folder anyway before going IDLE
-                // there, so this does not take additional protocol round-trip.
-                connection
-                    .fetch_move_delete(ctx, &mut session, &watch_folder, folder_meaning)
-                    .await
-                    .context("fetch_move_delete after scan_folders")?;
-            }
-            Ok(false) => {}
-        }
-    }
-
     // Synchronize Seen flags.
     session
         .sync_seen_flags(ctx, &watch_folder)
